Banking flow

+++ apps/dashboard/src/actions/accept-invite-action.ts
@@ -1,52 +1,51 @@
 "use server";
 
 import { LogEvents } from "@midday/events/events";
-import { setupAnalytics } from "@midday/events/server";
-import { getUser } from "@midday/supabase/cached-queries";
-import { createClient } from "@midday/supabase/server";
 import {
   revalidatePath as revalidatePathFunc,
   revalidateTag,
 } from "next/cache";
-import { action } from "./safe-action";
+import { authActionClient } from "./safe-action";
 import { acceptInviteSchema } from "./schema";
 
-export const acceptInviteAction = action(
-  acceptInviteSchema,
-  async ({ id, revalidatePath }) => {
-    const supabase = createClient();
-    const user = await getUser();
-
-    const { data: inviteData } = await supabase
-      .from("user_invites")
-      .select("*")
-      .eq("id", id)
-      .single();
-
-    await supabase.from("users_on_team").insert({
-      user_id: user.data.id,
-      role: inviteData.role,
-      team_id: inviteData.team_id,
-    });
-
-    await supabase.from("user_invites").delete().eq("id", id);
-
-    if (revalidatePath) {
-      revalidatePathFunc(revalidatePath);
-    }
-
-    revalidateTag(`teams_${user.data.id}`);
-
-    const analytics = await setupAnalytics({
-      userId: user.data.id,
-      fullName: user.data.full_name,
-    });
-
-    analytics.track({
+export const acceptInviteAction = authActionClient
+  .schema(acceptInviteSchema)
+  .metadata({
+    name: "accept-invite",
+    track: {
       event: LogEvents.AcceptInvite.name,
       channel: LogEvents.AcceptInvite.channel,
-    });
+    },
+  })
+  .action(
+    async ({
+      parsedInput: { id, revalidatePath },
+      ctx: { user, supabase },
+    }) => {
+      const { data: inviteData } = await supabase
+        .from("user_invites")
+        .select("*")
+        .eq("id", id)
+        .single();
 
-    return id;
-  }
-);
+      if (!inviteData) {
+        return;
+      }
+
+      await supabase.from("users_on_team").insert({
+        user_id: user.id,
+        role: inviteData.role,
+        team_id: user.team_id as string,
+      });
+
+      await supabase.from("user_invites").delete().eq("id", id);
+
+      if (revalidatePath) {
+        revalidatePathFunc(revalidatePath);
+      }
+
+      revalidateTag(`teams_${user.id}`);
+
+      return id;
+    },
+  );
